move assertions endpoint impls to their own file
